style: move hooks to hooks folder
