+++ wondrous-app/hooks/useCanManage.ts
@@ -2,7 +2,7 @@ import { PERMISSIONS } from 'utils/constants';
 import { parseUserPermissionContext } from 'utils/helpers';
 import { useBoards, useCreateEntityContext } from 'utils/hooks';
 
-export const useCanManage = () => {
+const useCanManage = () => {
   const { board } = useBoards();
   const { userPermissionsContext } = useCreateEntityContext();
   const permissions = parseUserPermissionContext({
@@ -17,14 +17,4 @@ export const useCanManage = () => {
   return canManage;
 };
 
-export const useCanEdit = () => {
-  const { board } = useBoards();
-  const { userPermissionsContext } = useCreateEntityContext();
-  const permissions = parseUserPermissionContext({
-    userPermissionsContext,
-    orgId: board?.orgId,
-    podId: board?.podId,
-  });
-  const canEdit = permissions.includes(PERMISSIONS.FULL_ACCESS);
-  return canEdit;
-};
+export default useCanManage;
